feat(cookie): expose splitCookiesString from h3
