+++ src/utils/session.ts
@@ -202,7 +202,10 @@ export async function unsealSession(
   return unsealed;
 }
 
-export async function clearSession(event: H3Event, config: Partial<SessionConfig>) {
+export async function clearSession(
+  event: H3Event,
+  config: Partial<SessionConfig>
+) {
   const sessionName = config.name || DEFAULT_NAME;
   if (event.context.sessions?.[sessionName]) {
     delete event.context.sessions![sessionName];
